+++ ASP.NETWebAPI/Matrices.json
@@ -4,12 +4,124 @@
       "MatrixName": "A",
       "Value": [
         [
-          5,
+          7,
+          4,
+          4,
+          7,
+          4,
+          7,
+          7,
+          9,
+          0,
           6
         ],
         [
+          9,
           2,
+          0,
+          6,
+          4,
+          5,
+          3,
+          7,
+          0,
           4
+        ],
+        [
+          0,
+          7,
+          5,
+          1,
+          0,
+          9,
+          5,
+          8,
+          2,
+          3
+        ],
+        [
+          9,
+          8,
+          4,
+          2,
+          9,
+          1,
+          8,
+          5,
+          5,
+          7
+        ],
+        [
+          7,
+          8,
+          3,
+          6,
+          2,
+          0,
+          3,
+          2,
+          0,
+          1
+        ],
+        [
+          5,
+          4,
+          7,
+          0,
+          4,
+          7,
+          5,
+          8,
+          3,
+          0
+        ],
+        [
+          7,
+          2,
+          7,
+          3,
+          8,
+          1,
+          8,
+          2,
+          8,
+          5
+        ],
+        [
+          1,
+          7,
+          5,
+          2,
+          6,
+          5,
+          9,
+          3,
+          8,
+          0
+        ],
+        [
+          3,
+          8,
+          7,
+          8,
+          9,
+          1,
+          9,
+          3,
+          5,
+          8
+        ],
+        [
+          0,
+          3,
+          0,
+          5,
+          6,
+          8,
+          4,
+          8,
+          4,
+          7
         ]
       ]
     },
@@ -18,11 +130,123 @@
       "Value": [
         [
           5,
-          8
+          7,
+          2,
+          6,
+          2,
+          9,
+          5,
+          6,
+          3,
+          0
+        ],
+        [
+          2,
+          0,
+          8,
+          1,
+          7,
+          1,
+          3,
+          7,
+          9,
+          2
+        ],
+        [
+          8,
+          5,
+          5,
+          3,
+          1,
+          6,
+          0,
+          2,
+          5,
+          2
+        ],
+        [
+          5,
+          2,
+          2,
+          4,
+          2,
+          8,
+          4,
+          5,
+          3,
+          7
         ],
         [
           1,
+          5,
+          8,
+          2,
+          7,
+          6,
+          1,
+          1,
+          2,
+          3
+        ],
+        [
+          8,
+          1,
+          8,
+          7,
+          9,
+          8,
+          4,
+          4,
+          6,
           8
+        ],
+        [
+          8,
+          0,
+          8,
+          6,
+          6,
+          2,
+          6,
+          9,
+          2,
+          1
+        ],
+        [
+          5,
+          4,
+          0,
+          6,
+          9,
+          2,
+          7,
+          0,
+          2,
+          4
+        ],
+        [
+          3,
+          0,
+          6,
+          4,
+          9,
+          5,
+          2,
+          5,
+          7,
+          7
+        ],
+        [
+          4,
+          7,
+          5,
+          8,
+          5,
+          6,
+          1,
+          4,
+          5,
+          7
         ]
       ]
     }
