+++ ASP.NETWebAPI/Program.cs
@@ -1,3 +1,6 @@
+using ASP.NETWebAPI.Services;
+using ASP.NETWebAPI.Services.Interfaces;
+
 var builder = WebApplication.CreateBuilder(args);
 
 // Add services to the container.
@@ -6,6 +9,7 @@ builder.Services.AddControllers();
 // Learn more about configuring Swagger/OpenAPI at https://aka.ms/aspnetcore/swashbuckle
 builder.Services.AddEndpointsApiExplorer();
 builder.Services.AddSwaggerGen();
+builder.Services.AddScoped<INumbersService, NumbersService>();
 
 var app = builder.Build();
